% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_prism.R
\name{model_run}
\alias{model_run}
\title{Run CVDANDERSON Model}
\usage{
model_run(model_input = NULL)
}
\arguments{
\item{model_input}{A list/json object with "gender", "age",
"Tchol","HDLchol","SBP", "DBP", "diabetes",
"smoker" , "ECG_LVH" , "t"}
}
\value{
Return a dataset with three columns: risk_Outcome, systolicbase_risk, and diastolicbase_risk
}
\description{
This function calls the CVDANDERSON model.
}
